requirements.txt
streamlit
numpy
gspread
gspread-dataframe
google-auth-oauthlib
